Add basic aarch64 hello world
This code compiles into an aarch64 ELF binary that when loaded just
sends out a hello world message via UART.

UART device I use in this case is PL011, which is supported by QEMU virt
and in general a rather common UART implementation in the ARM world,
since PL011 was designed by ARM itself.

Here is the command I use to test it in QEMU:

qemu-system-aarch64 -machine virt -cpu max
  -drive if=pflash,format=raw,file=${OVMF} \
  -drive format=raw,file=fat:rw:/home/kmu/ws/aarch64/root \
  -net none \
  -serial telnet:localhost:1234,server \
  -monitor telnet:localhost:1235,server,nowait \
  -nographic

where

 * OVMF points to the image with UEFI firmware for QEMU aarch64
 * /home/kmu/ws/aarch64/root is the path to my root file system which is
   the place where I put the compiled binary as well as the loader
   binary

The UART output can be observed via telnet on the localhost:1234. QEMU
monitor can be interacted with via telnet on the localhost:1235.

The EFI loader binary I use can be build from the sources in
git@github.com:krinkinmu/efi.git (https://github.com/krinkinmu/efi.git
for https). The binary assumes that the kernel image is ELF and is
stored in the efi/boot/kernel file.

The clock frequency and the base address for the PL011 registers memory
block can be found in the DTB (Device Tree Blob) for the QEMU virt
platform. Here is how you can get the DTB (output will be placed in the
virt.dtb):

qemu-system-aarch64 -machine virt,dumpdtb=virt.dtb -cpu max

The output will be in the binary format, so you can use dtc (Device Tree
Compiler) to convert it into DTS (text format).