Add "support" for HiKey960
HiKey960 board also uses PL011 based UART. The "support" here means a
comment that shows the parameters that should be used on the board plus
some fixes that I had to make after testing it on HiKey960 (see below).

The most important part I had to fix for this code to work on HiKey960
is waiting for outgoing transfers to complete before doing the reset of
the device.

I'm not entirely sure what is causing this problem, but I know that
previous writes to the UART unless completed cause a problem. In my case
previous writes were generated by the UEFI firmware. Evidently UEFI
firmware doesn't wait for all outgoing transfers to complete even when
we exit boot services. Adding a wait on the outgoing transfers masked
the problem.

I cannot claim that it addresses the problem completely, because I don't
really know what interaction causes it.

In the course of my experiments I also introduced a write to the ICR
register because I was supsecting that interrupts might be somehow
involved into the problem. While it didn't actually resovle the problem
I left it in the code, since it doesn't cause any harm and puts the
device into a better defined state.