Add hacky Rust building infrastructure to the project
The idea of how it all is supposed to work together is as follows:

 * entry point is not written in Rust (it's currently in C) and we can
   do whatever C/assembly stuff we want in the entry code

 * eventually the entry point calls into Rust code compiled as a static
   library

As a PoC I implemented PL011 code in Rust as well.

I'm not a big fun of Rust, but there is at least one benefit to explore
there and it's integrated support for tests.

Unfortunately, when you start thinking a little bit deeper about tests
for cross-compiled projects there are quite a few configurations for
tests that we might want to consider:

 * tests running on the target platform
 * tests running on the host platform.

Both configurations make sense and it's likely that in a mature enough
project we want to have both.

Here I tried to explore only the first option to create unit tests that
do not depend on low level platform details (like unit tests for common
data structure or basic computations). However even that is somewhat
complicated.

For example, it requires support for two different target architectures
(aarch64 for the main binary plus whatever host platform is used for
building and testing).

Another example, is that not all crates make sense for building both
tests and the final binary. Specifically, Rust bare metal binaries need
some minimal runtime that not only makes sense for tests, but in fact
causes link conflicts when building tests (since for test that runtime
is already provided).

All-in-all, it works somehow, but the setup is far from perfect.