Reorganize the code to separate bootstrap logic from kernel
There are a few things that this change does:

 * puts the bootstrap code into its own directory
 * renames libstart.a to libkernel.a

The goal of bootstrap code is to collect all the required information
from the bootloader or by whatever means necessary and wrap it into the
interface that the kernel will understand.

Currently the interface is defined just as a structure that contains a
pointer to the DTB and the list of reserved memory ranges. We can add
more things in the future if needed.

Renaming libstart.a to libkernel.a makes sense because that's basically
what libstart.a - it's our kernel at this point, even though an
extremelly simplistic one.

I adjusted the build system a little bit as well to build the bootstrap
code into a static library.