Add basic device tree defintions
This change doesn't contain any actual device tree parsing logic (that
part will come later). At this point it only includes basic interfaces
and build/test infrastructure.

Parsing/traversing the device tree will be implemented using a visitor
pattern. It's somewhat inconvenient to use, but the reason to do it this
way is that we don't have dynamic allocation available, so we sort of
limited to working with the device tree in-place.

Note that I'm going to implement dynamic allocation, but the information
provided via device tree contains some inputs that I need to setup the
dynamic allocation. Specifically, it contain the description of the
available memory in the system and we need to know that to organize the
dynamic memory allocation. As a result we need to be able to parse the
device tree without dynamic memory allocation available.