Plug in the DeviceTree into the start_kernel function
To make sure that the devicetree library compiles correctly this change
adds a call into the devicetree library from the start_kernel function.

Along the way I added the BootInfo structure description. This
structure is passed from the bootstrap code and includes pointers to the
device tree blob in memory.