Implementation of DeviceTree parsing with tests
That's a lot of code, but most of the code is in tests (probably because
the implementation and the interface aren't perfect, but we can improve
on that further down the road).

The implementation is split in a few separates pieces:

 * scanner.rs - implementation of deserialization of basic primitives
   from the byte stream
 * visit.rs - definition of the visitor trait
 * lib.rs - the actual DeviceTree traversal.

The DeviceTree parsing was implemented as a visitor pattern, so the
parsing code will tigger callbacks when it discover certain kinds of
objects in the DeviceTree.

Alternative implementation of parsing might define types to represent
the DeviceTree nodes and their properties and return them after parsing
instead of triggering callbacks.

The alternative implementation is certainly more convenient to use, but
it's hard to return any kind of dynamically built structture without
dynamic allocation available.

And to make the dynamic allocation available we need to figure out the
memory structure first. And as you may have guessed it, to obtain this
information we need to parse the DeviceTree first. Thus I went with a
less convenient interface that doesn't require dynamic allocation.

One additional note, test.dtb file is used in tests, that's why it's
part of this repository. The test.dtb fail was generated from the
test.dts file (also part of this change) using the DeviceTree comapiler
as follows:

$ dtc -I dts -O dtb -o test.dtb test.dts