A simplistic dynamic memory allocator
After dealing with the DeviceTree I realized that it's quite messy to
try to work with rather complicated configuration frameworks without
dynamic memory allocation and as a result the interface of the
DeviceTree library that I created wasn't particularly convenient.

So I decided to go a slightly different way and provide a simplistic
dynamic memory allocatior that could be used during the early stages
without knowledge of the all the available memory in the system.

This allocator will be handy to complete all the setup and get the
memory map needed for the final memory subsystem. Once the final memory
subsystem is initialized we can free all the memory and shutdown the
early memory allocator.

The implementation itself is quite simple: it's just linked lists of
free blocks in the memory that the allocator distributes. I use a known
trick of boundary tags to quickly find adjusent memory ranges and check
if they are free or not on deallocation to avoid fragmentation.

It's exepected that the memory range that this allocator will use will
be statically allocated (as a global variable in the binary).