Introduce a new configuration interface
I previously was collecting the configuration inside a structure, but
it's more convenient to provide a set of functions instead to obtain the
configuration information.

Essentially those functions are global, so it's slightly easier to use
them compared to passing a configuration structure everywhere.

This change only introduces the interface, but it's not actually plugged
anywhere.