Mask all the IRQs and FIQs on the boot CPU early
I noticed in the QEMU monitor that there are some periodic intrrupts
happenning in the system when running under QEMU. So I suspect that
there might be some interrupt handler in memory and things can go south
if we decide to overwrite the memory where the handler is stored.

It's safer to just mask all the external interrupts until we figure out
what devices are there and setup the required handlers.