Install a dummy isr vector table and normalize the environment
The dummy ISR vector table contains entries that just go into an
infinite loop when they are called, since at the current moment
there shouldn't be anything that requires processing there.

As for normalization of the environment, it means that I tried to
settle on some basic CPU configuration:

 - non-secure EL2
 - MMU is off
 - all maskable interrupts are masked
 - SPSel = 1 (each EL uses dedicated SP register)
 - custom ISR vector table is configured

And make the startup code make necessary changes to set the required
configuration if possible (or fail if it's not possible).

I couldn't test the code that jumps from EL3 to EL2 though, since the
OVMF firmware that I have for some reason doesn't work in EL3 in QEMU.
I don't know what the issue there yet, but either way that part of the
code has not been tested and cannot be trusted to work. The other parts
work at least in QEMU and tests on a real hardware are to be done.