Disable 16-byte alignment checks for SP
It's not clear to me what mechanisms prevent taking an exception or an
interrupt with a misaligned value in SP. Given that when we are taking
an interrupt we have to store the values of at least of GPRs on stack
it's rather tricky to handle it if you also need to be worried that the
value in SP might be potentially misalgned.

It's easy to circumvent the alignment problem if we use another register
instead of SP as a base for memory accesses, however then it creates a
problem of saving that register value somewhere and since we cannot use
stack not many options are available.

So it's easier to just disable the stack alignment checking to be able
to preserve the old value of SP on stack and then just align the SP to
the 16-byte boundary manually to still stick to the 16-byte alignment
everywhere else except the ISR entry.