Introduce basic interrupt handling logic
The current logic just calls into a C routine (doing all the normal
stuff like preserving GPRs and so on) and doesn't actually handle
anything. Moreover as of now we don't even provide enough information
to the handler to understand the cause to begin with.

However the current logic is enough to verify that our
interrupt/exception setup works in general and that the GPRs and other
registers we want to preserve are preserved across the interrupt
handlers. So that's good.

I tested the logic by generating a data about by dereferencing an
incorrect pointer and then comparing the values in the registers before
and after the handler using QEMU monitor (info registers) and GDB (to
set breakpoints).