Introduce basic interrupt/exception handlers in Rust
They aren't doint much, just print the saved registers to the PL011
serial port.

Note that using serial port this way is quite problematic in general. We
create another instance of PL011 structure that uses the same hardware
as the one we currently have in start_kernel. So we basically accessing
shared hardware resources from multiple places in code without any
synchronization.

Rust type-system is completely useless here and the only way to protect
from this kind of problems is via correct logic and runtime checks.
Currently we don't need them as I'm just playing around and that's not
how exceptions and interrupts need to be handled in general.