Rework the DeviceTree library interface
Previous version of the parser was callback based in the attempt to
avoid allocating memory dynamically. So as we go through the FDT the
parser was calling various callbacks with references to the data in the
FDT.

The interface was however somewhat terrible in a few ways:

* it didn't actually solve the problem of avoiding dynamic allocations,
  but instead pushed it on the user to solve (which you can see if you
  look at the tests of the previous version)
* the user of the library through the interface was basically exposed to
  the internal encoding of the FDT, as the callbacks were 1:1 mapped to
  the FDT tokens
* every user had to write a bunch of somewhat obscure code to be able to
  use the Device Tree library.

This change reworks the Device Tree to parse the FDT and return a some
kind of a tree structure describing the content of the FDT. So instead
of defining callbacks the users of the library can just traverse the
returned tree structure using iterators and provided helper functions
(though there aren't many at the point).

One downside at the moment is that the previous version had tests and
this one doesn't. I'm planning to fix that soon though.