Extract FDT parsing logic into a separate module and add tests
This change finishes my rework of the Device Tree parsing logic and
interface. I may make some more changes in the future when I start using
the library, but those should be more or less minor unless I overlooked
something.

I extracteed the FDT parsing logic into a separate module because lib.rs
was getting cramped and I was also contemplating adding other parsers
(for example a DTS parser) as well. So the structure where parsers are
in a separate module makes slightly more sense.

The only public interface function of the fdt module is parse function
that takes the FDT and returns the DeviceTree representation parsed from
that FDT (or error if FDT was incorrect).