Move global allocator to runtime and introduce memory crate
As far as user is concerned Rust global allocator is a runtime as it's
required by the Rust language. And yes, Rust core library is part of the
Rust language and it does require the global allocator unfortunately.

I remove the memalloc create as it only hosted the global allocator, so
there is no need in it anymore.

Finally, I'm introducing a new create called memory. I'm going to put
there my code for working with memory. The current plan is to put there
the implementation of a few custom memory allocators.

At this point the code in the memory crate contains a fairly silly
linked list. That's not completely random, as I will need some form of a
linked list in one of the algorithms that I'm planning to implement,
but, all-in-all, the existing code in the create is far from what I
intend to implement there. So consider it just a crate structure change.