Implement Buddy allocator and tests for it
The implementation consists of a few pieces:

* Page structure - contains the book keeping information per page frame
* PageRange structure - basically exists to map from Page structures to
  their indexes and vice versa
* BuddySystem - the actual allocator, that manages the list of free
  pages to be able to effectively allocate contigous ranges of page
  frames that are 2^n page frames long

Algorithm doesn't really care what struct Page represents. It can
represent a memory range, space on disk or really anything. However in
this case what I intend to use this system to allocate memory.

So struct Page represents a memory range of a fixed size - page frame.
You can think that all memory is partitioned in an array of page frames.
And for each page frame we have a struct Page. Index of the struct Page
is just a position of the page frame it represents in the array.

To allocate or free page frames we need both BuddySystem instance and
the corresponding PageRange. So I additionally introduced a struct
called Zone which pairs them together. So you can think of Zone as a
manager of a contigous memory range.

The memory overall doesn't have to be contigous. And while we can
represent the whole memory with just one Zone (and treat holes in memory
as if they are always busy) I decided to support multiple zones.

All the Zone structures are combined inside Memory structure that
represents the whole memory. Memory structure also serves as an entry
point to the API. All other structures introduced in the change are not
public, so the user should not be able to interact with them.

This change doesn't include the code to initialize Memory structure as
it's a complicated problem in its own right. The change does include
tests for the Buddy allocator though, so I have some level of confidence
that it's not completely broken, even though it's not complete.