Introduce the code initializing memory
The code expected the caller to provide a sort of memory map that
contains the list of memory ranges available in general as well as the
list of reserved memory ranges.

The code creates a zone for each contigous range of memory provided.
TO store the Page structs for the zone the code tries to find a free
memory range within the zone large enough to hold all the Page structs
needed to describe it.

If there is not enough free space in the zone to hold the Page structs
we declare the zone unusable. It's a bit of a problem, because if
somebody will try to later free some space in the zone the operation
will fail. However I for now don't want to deal with that problem and
would say that it's somewhat unlikely to happen. Nonetheless it's still
a bug.

All pages within the zone are zero initialized. Rust doesn't really have
a concept of zero-initialization, so in general there is no guarantee
that a particular type can be zero-initialized. However Rust does
provide an intrinsic function that could be used to assert that a
particular type is zero-initializable. I use that intrinsic to assert
that Page struct is zero-initializeable.

Since all the pages are zero initialize it means that they all marked as
busy level 0 pages. So all the memory is considered occupied by the
Buddy allocator. I will fix it in the subsequent changes.