Add a sync library
The library is not very useful at this stage and I only created a
placeholder Mutex type. It doesn't actually provide any mutual
exclusion, but it does provide the same type properties that a normal
mutual exclusion primitive should provide.

It's obviously unsafe, as the only reason the normal Rust Mutex is safe
is because of its runtime guarantees, but it's hard to work on
everything at the same time, so this placeholder Mutex, even though
unsafe, is quite useful.