Fix memory module bugs and introduce more tests
This time I actually manually run some tests against the memory module
and, no surprise, it was quite buggy. To begin with the memory map
module code for reserving memory ranges wasn't correct and didn't handle
the situation when we reserve a range that lies completely within
another range well.