Refactor buddy allocator and fix bugs
This changes gets rid of PageRange abstraction and replaces linked list
on indexes with linked list on pointers. This makes the code
structurally simpler because there is less dependencies between
individual submodules (we don't need to pass PageRange around) and less
entities to deal with.

The responsibilities of PageRange either aren't needed anymore (because
the list of pages can now directly dereference the pointers instead of
converting indecies to pages) or take on by BuddySystem.

While working on this refactoring I also realized that when freeing
nodes BuddySystem only checked if the buddy is free, but not whether
it's on the right level, which was incorrect. I fixed the implementation
and added a test case covering this situation.