Replace *const Page with &Page in list.rs
It doesn't make the list safe, because list links still can be modified
outside of the list and corrupt everything (because of the interior
mutability).

However it does allow to remove a few asserts checking for null
pointers, since references cannot be null. Plus it allows to avoid
constant conversion between *const Page and &Page in the client code by
hiding those conversions inside the list.rs.

As a side bonus it puts some constraints on liftimes of List and Page.
It doesn't matter since all Page structures will have 'static lifetime
and the code is unsafe anyways and Rust cannot check much. However at
least it can serve as documentation.

Finally, note that links to the next and prev elements inside the Page
structure are still expressed as pointers. The reason for that is that
we need Page structure to be zero-initializable (initializing 10^6s of
pages one by one is kind of costly and even with moderate 4G of memory
we will already have 1000000 pages to initialize), &Page is certainly
not zero-initializable and it's not obvious if Option<&Page> is
zero-initializable. So we have to stick with pointers there instead of
Option<&Page> that would normally work.