Implement generic intrusive linked list
I found myself implementing lists over and over again. This is my
attempt to extract at least some logic into a common library. This
implementation is naturally unsafe.

I'm probably not going to use this implementation directly and where I
need a linked list I'll to wrap this implementation to try to provide a
somewhat safer interface. So this library is purely to avoid code
duplication and not to provide a generic container.