Get rid of global logger
It turned out to be quite inconvenient to use due to the way the logger
has to be initialized.

When running in VM/on hardware I use serial port and when running tests
I don't want to use logs at all or use stdout/stderr. And turns out
making sure that the global logger is initialized the way I want
depending on the environment is a pain.

So I'm removing it for now until I figure out how to make it convenient.