A rework of the aarch64 codebase
This particular change is very much work in progress that was sitting
on my laptop for a while without much progress. I'm commiting it just
to avoid losing it at some point.

Roughly the changes replace C and Rust with C++. Naturally, I'm not a
big fan of Rust, but objectively allocators, intrusive list and
similar low-level hacky memory manipulations that I do in the code are
not the Rust way of doing things, so for a fun project, it does not
make sense to struggle through it.

I might return back to Rust, once I have basic runtime libraries ready.
Once those are availbe, creating Rust bindings and implementing Rust
runtime should be a piece of cake and once that is done, we can just
use proper Rust instead of whatever I was using.