Add a basic OutputStream interface
It's a very simplistic OutputStream interface. I intend to use it my logging
code. Currently, the logging code lives in the bootstrap library and combines
a few things:

  * interraction with the actual device we write logs to (PL011 in our case)
  * formatting and string manipulations.

I want to separate formatting and string manipulations from the interaction
with the serial interface. The OutputStream interface is what the code that will
do formatting and string manipulations will interact with.

The bootstrap code will have to implement OutputStream interface on top of the
PL011 device and register it with the logging code.