Expose Zone as public interface
There are cases where I don't want to store the whole Contigous structure
(which includes Page pointer, order and Zone pointer) and want just an address.

Often I know the order just because it's a fixed number (for example if some
code always allocates single pages or a fixed number of pages), so the only
thing I need to infer to free allocated memory is the Zone. Once I know the
Zone and the address, I can easily find the Page.

For example, when working with virtual memory code, I only need to allocate
chunks of a fixed known size (4kb or 64kb depending on the configuration). The
page table structure naturally already stores an address. If we had to store
order and Zone pointer somewhere as well, we would need to allocate and manage
some additional memory, which is not convenient.