Drop unque_ptr from the buddy allocator interface
It looks natural to wrap allocated memory into a smart pointer, but in
practice I do not use the unique_ptr anywhere. The reason for that is
that I cannot use Contigous as a regular memory pointer, as a result the
only property of the unique_ptr that is still relevant is automatic
release of the object, but there is just no code that needs it at the
moment.